voice-agent/pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Voice agent for camera control"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "pyaudio>=0.2.11",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_agent"]

